[feature] rename class name
