+++ schema-engine-tag/src/main/java/org/apache/iotdb/db/metadata/tagSchemaRegion/tagIndex/query/MemTableQuery.java
@@ -19,8 +19,8 @@
 package org.apache.iotdb.db.metadata.tagSchemaRegion.tagIndex.query;
 
 import org.apache.iotdb.db.metadata.tagSchemaRegion.tagIndex.Request.QueryRequest;
-import org.apache.iotdb.db.metadata.tagSchemaRegion.tagIndex.memtable.MemChunkGroup;
 import org.apache.iotdb.db.metadata.tagSchemaRegion.tagIndex.memtable.MemTable;
+import org.apache.iotdb.db.metadata.tagSchemaRegion.tagIndex.memtable.MemTagValueGroup;
 import org.apache.iotdb.lsm.annotation.QueryProcessor;
 import org.apache.iotdb.lsm.context.requestcontext.QueryRequestContext;
 import org.apache.iotdb.lsm.levelProcess.QueryLevelProcessor;
@@ -33,23 +33,23 @@ import java.util.Set;
 
 /** query for MemTable */
 @QueryProcessor(level = 1)
-public class MemTableQuery extends QueryLevelProcessor<MemTable, MemChunkGroup, QueryRequest> {
+public class MemTableQuery extends QueryLevelProcessor<MemTable, MemTagValueGroup, QueryRequest> {
 
   /**
-   * get all MemChunkGroups that need to be processed in the current MemTable
+   * get all MemTagValueGroups that need to be processed in the current MemTable
    *
    * @param memNode memory node
    * @param context request context
    * @return A list of saved MemChunkGroups
    */
   @Override
-  public List<MemChunkGroup> getChildren(
+  public List<MemTagValueGroup> getChildren(
       MemTable memNode, QueryRequest queryRequest, QueryRequestContext context) {
-    List<MemChunkGroup> memChunkGroups = new ArrayList<>();
+    List<MemTagValueGroup> memTagValueGroups = new ArrayList<>();
     String tagKey = queryRequest.getKey(context);
-    MemChunkGroup child = memNode.get(tagKey);
-    if (child != null) memChunkGroups.add(child);
-    return memChunkGroups;
+    MemTagValueGroup child = memNode.get(tagKey);
+    if (child != null) memTagValueGroups.add(child);
+    return memTagValueGroups;
   }
 
   /**
